Borewit/music-metadata-browser#5: Restructure to centralize types and core (common Node & Web) functions.
